+++ src/main/java/com/carrental/requestmodel/ExtraFeeRequest.java
@@ -2,12 +2,10 @@ package com.carrental.requestmodel;
 
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 public class ExtraFeeRequest {
     private String name;
     private Long limit;//limit distance,...
     private String unit;//km,vnd,..
-    private BigInteger fee;
+    private Double fee;
 }
