+++ src/main/java/com/carrental/responsemodel/ExtraFeeResponse.java
@@ -2,13 +2,11 @@ package com.carrental.responsemodel;
 
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 public class ExtraFeeResponse {
     private String name;
     private Long limitValue;
     private String unit;
-    private BigInteger fee;
+    private Double fee;
     private String code;
 }
