+++ src/main/java/com/carrental/service/IExtraFeeService.java
@@ -2,11 +2,10 @@ package com.carrental.service;
 
 import com.carrental.entity.ExtraFeeEntity;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public interface IExtraFeeService {
     List<ExtraFeeEntity> saveAll(List<ExtraFeeEntity> list);
 
-    BigInteger findDeliveryToTenantFee(Long id);
+    Double findDeliveryToTenantFee(Long id);
 }
